Refactor: Make folders for daejeon
