+++ scrap/local_councils/__init__.py
@@ -2,5 +2,5 @@
 각 기초의회들의 크롤링 코드를 모아놓은 패키지입니다.
 광역자치단체 별로 폴더를 만들어서 관리합니다.
 """
-from .daejeon import *
+from .daejeon.daejeon import *
 from .ulsan import *
